Integrated communication protocol with our motor controller
Tested this implementation against our unit tests and all tests came
across green!

Changed motor throttle pins so that pin 13 when on, indicates that the
device is in its ACTIVE state.

Moved motor config struct to the header file and added const accessors
to it so that they can be read out by telemetry if necessary.